Add files via upload
This Dir is best on understanding recommendation systems such as Amazon, ebay, or any organization that uses any service related to recommendation. Along with the critical there is a working recommendations system with training data . If your using newer hardware you should be OK  but on older hardware due to the complexly and work needed your computer will over head. RECOMMENDATION: using a public cloud service to run the Eigen3 based recommendation application with a smaller data set.