Add files via upload
I'm still working on the ensemble application finding and configuring all the dependencies took a lot long then planned today. In the mean while I uploaded something on neural networks. It the bases of of a project I'm going to try to how to a cloud provider to demonstrate the critical importance of distributed systems and cloud computing for ML and overall progress. That will lead to my Openstack hybrid cloud solutions for automating new upstarts to dusting off the old dirt on the legacy corporations that could use upgrades but while destroying tech debt.