Add files via upload
Hey there was random switch in the code I was going to post for that ensomble dir project which its still happening but I notice design issues on the new one and I'm like 90 percent sure there was a std segfault in the program. Also with no documentation of the new algorithm there was no discussion on  the 3rd party library (I dont like). That being said you have to go though DNN and CNN first before any  other neural network.