Add files via upload
This is some of the most complex things in tech and its not easy to understand. Your really learning how to analyze the thing that powers your phone to log into Facebook with just looking at your face. This is what amazon is in fight with Microsoft, google cloud, linnode, alibaba, and other companies. This is why your GPU is so valuable. I can't put into words that the knowledge I'm posting is, and the power that comes with understanding and advancing. Please don't not open this DIR if your not one of us.