Add files via upload
OK, Now you can look at this dir. my last upload, I didn't expect too many to get all the README. Albeit, now with the recent upload you can truly understand all the concepts of neural network and have a good grasp of RNN's. Do not read the commit comment before this I was based on for the point of view of someone trying to learn with no pre knowledge of biology and stats integration.